requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
requests
scikit-learn
python-dotenv
numpy
Jinja2
